Add files via upload
PROGRAM LIST 

1. Aim: Introduce the Python fundamentals, data types, operators, flow control and exception handling in Python 
a) Write a python program to find the best of two test average marks out of three test’s marks accepted from the user. 
b) Develop a Python program to check whether a given number is palindrome or not and also count the number of occurrences of each digit in the input number. 

2. Aim: Demonstrating creation of functions, passing parameters and return values 
a) Defined as a function F as Fn = Fn-1 + Fn-2. Write a Python program which accepts a value for N
 (Where N >0) as input and pass this value to the function. Display suitable error message if the condition for input value is not followed. 
b) Develop a python program to convert binary to decimal, octal to hexadecimal using functions. 

3. Aim: Demonstration of manipulation of strings using string methods 
a) Write a Python program that accepts a sentence and find the number of words, digits, uppercase letters and lowercase letters. 
b) Write a Python program to find the string similarity between two given strings 

4. Aim: Discuss different collections like list, tuple and dictionary 
a) Write a python program to implement insertion sort and merge sort using lists 
b) Write a program to convert roman numbers in to integer values using dictionaries. 

5. Aim: Demonstration of pattern recognition with and without using regular expressions 
a) Write a function called isphonenumber () to recognize a pattern 415-555-4242 without using regular expression and also write the code to recognize the same pattern using regular expression. 
b) Develop a python program that could search the text in a file for phone numbers (+919900889977) and email addresses (sample@gmail.com) 

6. Aim: Demonstration of reading, writing and organizing files. 
a) Write a python program to accept a file name from the user and perform the following operations 1. Display the first N line of the file 2. Find the frequency of occurrence of the word accepted from the user in the file 
b) Write a python program to create a ZIP file of a particular folder which contains several files inside it. 

7. Aim: Demonstration of the concepts of classes, methods, objects and inheritance 
a) By using the concept of inheritance write a python program to find the area of triangle, circle and rectangle. 
b) Write a python program by creating a class called Employee to store the details of Name, Employee_ID, Department and Salary, and implement a method to update salary of employees belonging to a given department. 

8. Aim: Demonstration of classes and methods with polymorphism and overriding 
a) Write a python program to find the whether the given input is palindrome or not (for both string and integer) using the concept of polymorphism and inheritance. 

9. Aim: Demonstration of working with excel spreadsheets and web scraping 
a) Write a python program to download the all XKCD comics 
b) Demonstrate python program to read the data from the spreadsheet and write the data in to the spreadsheet 

10. Aim: Demonstration of working with PDF, word and JSON files 
a) Write a python program to combine select pages from many PDFs 
b) Write a python program to fetch current weather data from the JSON file